refactor: runPgLocals to Query.hs
* unmmiddleware runPgLocals
